requirements.txt
asgiref
sqlalchemy
aiogram==3.1.1
aiohttp>=3.6.2
aioredis==1.3.1
python-dotenv~=1.0.0
Django==4.2
psycopg2-binary
jsonfield
cachetools~=5.3.2
redis~=5.0.1